+++ package.json
@@ -6,11 +6,15 @@
   "dependencies": {
     "@settlin/collage": "^1.4.3",
     "@types/node": "^12.0.12",
+    "child_process": "^1.0.2",
     "evolve-ga": "^2.0.1",
-    "typescript": "^3.5.2"
+    "express": "^4.17.1",
+    "typescript": "^3.5.3"
   },
   "devDependencies": {},
   "scripts": {
+    "build": "tsc",
+    "start": "node server.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "mobator": "tsc genetic-algorithm.ts -resolveJsonModule && node genetic-algorithm.js gank 2095",
     "hardengage": "tsc genetic-algorithm.ts -resolveJsonModule && node genetic-algorithm.js hardengage 2095",
